use semantic tags in selectors
The `.content` and `.hero-body` style classes are part of the CSS framework I happen to use, and I may change that framework in the future. A better set of selectors would leverage the `header` and `main` semantic elements, which I use in my template. I'm careful to include indexable content only within those elements.