Reconfigure Apollo to work across all docs repos.
Most all of the Apollo docs repositories now use the same format and I'm
hoping this will allow us to re-enable search on all of them, with a simple
configuration that will automatically expand as we deploy more properties.

A partial list of docs which would use this new configuration would be:

 * https://github.com/apollographql/apollo
 * https://github.com/apollographql/apollo-client
 * https://github.com/apollographql/apollo-link
 * https://github.com/apollographql/apollo-angular
 * https://github.com/apollographql/apollo-ios
 * https://github.com/apollographql/apollo-android
 * https://github.com/apollographql/graphql-tools
 * https://github.com/apollographql/apollo-server
 * https://github.com/apollographql/subscriptions-transport-ws
 * https://github.com/apollographql/engine-docs

It's possible that some of these haven't been updated yet, but they will be
soon!