Update uniwebview.json
We recreated our doc page on UniWebView: http://docs.uniwebview.com/

I am not quite sure about the config file since initially, some Algolia staff helped me to config the site (thank you a lot!)

Previously it is not a static page but viewers have to render it in the browser (so the js render was used). Now it is a pure static website so I guess I need to change something. 

There is some tricky on the API reference (like on this page http://docs.uniwebview.com/api/uniwebviewmessage.html)