Adding Phalcon
I'm having trouble selecting the correct lvl1. The element in the
sidebar is only bolded through style and there is no class attached
to it. Maybe we should ask the owner to add one?

Also I hard-coded the versions, maybe I should have used the JS
snippet to get the versions, but that seems brittle.