Update activepivot.json
# Pull request motivation(s)

### What is the expected behaviour?
New versions should be crawled as they are added to the [sitemap index](https://www.activeviam.com/activepivot/sitemap.xml). The json file needs to be cleaned, hence this PR. 
Please note the documentation website for version 5.8-next is not crawled, pages are not indexed (could it be related to https://github.com/algolia/docsearch-configs/pull/698#issuecomment-456520752). It seems to be a bug. Please advise.   

Thanks.