Remove directory path
So, when I enable the search, my results are linking to:

`http://masstransit-project.com/MassTransit/MassTransit/getting-started.html`

Which is a double path link. The generated content for VuePress is in /MassTransit, so I'm taking off the extension to hopefully? resolve the issue. If there is another setting to fix it, I haven't found it yet.