Update chromatic.json
Would like to update DocSearch to reflect moving our docs from a subdomain to a path. Would like to prevent the crawler from indexing the marketing site outside of the /docs/ subpath and have included a `stop_urls` entry to do so. This is my first time configuring DocSearch so I would appreciate it if you let me know if I got the syntax wrong. Thank you!