Added 'location' as an attribute for faceting
Hi Sylvain,
After the last changes, I could see the search results were populated. Thanks.
I want to try facet filtering on an attribute, "location". I've added a meta tag to each page:
<meta name="docsearch:location" content="coventry-university-group,cu-coventry,coventry-university">
I'll need to add this to the facetFilters in the search component - but is there anything else I need to do so that the crawler gets this?
Thanks
Rich