config: add top level selector as a facet
At Elastic Path, we are building a custom Search UI for our docs
site that is integrated with Algolia DocSearch service. In our
custom search UI, we need to add an option to be able to filter
search results based on results "Category". This "Category" is
same or equivalent of `lvl0` selector.

Added the `lvl0` selector as one of the facets.