Added new start_url to Twenty57_linx config
The new start_url contains a full menu.
We have optimised the menu on our site and now use JS to load only what needs to be shown in the menu. 
The indexer was therefor missing pages as they were not in the menu. The new start_url contains all pages in the menu.