Update config for playwright
We are building a new version of the Playwright docs website with Docusaurus v2. This PR updates the config to point to the [v2 preview website](https://microsoft.github.io/playwright/) and updates the config for Docusaurus v2, building on [this config](https://github.com/algolia/docsearch-configs/blob/master/configs/docusaurus-2.json).