update pype.json config
Hi.

We've recently updated from docusaurus 1 to v2, and our search stopped working. After looking through I suppose it could be due to minor changes in the structure. I'm changing the start_url and removing languages and version as those are no longer needed. 

Hopefully it will be able to crawl through the web again, with this.

thank you