allow Gatsby to filter by hierarchy
Similar to what is described in this discourse thread (https://discourse.algolia.com/t/how-to-configure-a-new-facet/10776/3), we want to filter by hierarchy, eg, only show results from the docs when searching in the docs. It sounds like to do this we need to add this custom_settings key.

Update gatsbyjs.json

Update gatsbyjs.json

fix