Update health-data-hub-snds.json
We have a lot of pages in our documentation. Therefore, the primary role of the search bar for us is to find a page and not a specific section within the page.
Currently, the search results display too many results per page and other pages are not shown. Ideally, we would like search results to only show one occurence for each page. 

I hope this change will be enough for the intended purpose.