Update objectiv.json
The pages of the Modeling section in our docs (https://objectiv.io/docs/modeling/) load content dynamically. 

This PR adds the `js_render` and `js_wait` options to our config to enable crawling the content on those pages. We set `js_wait` to 2 seconds to be on the safe side, though the pages generally load in a few hundred milliseconds; we hope the impact on crawling speed is not too high.